migrate to double

+++ Lab_3_2/src/lab32/model/AbstractForm.java
@@ -11,7 +11,7 @@ public abstract class AbstractForm implements IWeight {
         return wood;
     }
 
-    public float weight() {
+    public double weight() {
         return wood.getDensity() * volume();
     }
